+++ src/variables/aboutConsts.js
@@ -1,8 +1,8 @@
 const GLOBAL_NAME = "BubbleOS";
 const AUTHOR = "Arnav Thorat";
 
-const VERSION = "0.5.8";
-const BUILD = "58";
+const VERSION = "0.6.0";
+const BUILD = "60";
 const IN_BETA = true;
 
 export { GLOBAL_NAME, AUTHOR, VERSION, BUILD, IN_BETA };